Moving towards a pypi package structure for adding integration tests
