+++ deepspeech_pytorch/inference.py
@@ -1,0 +1,13 @@
+import torch
+
+
+def transcribe(audio_path, spect_parser, model, decoder, device, use_half):
+    spect = spect_parser.parse_audio(audio_path).contiguous()
+    spect = spect.view(1, 1, spect.size(0), spect.size(1))
+    spect = spect.to(device)
+    if use_half:
+        spect = spect.half()
+    input_sizes = torch.IntTensor([spect.size(3)]).int()
+    out, output_sizes = model(spect, input_sizes)
+    decoded_output, decoded_offsets = decoder.decode(out, output_sizes)
+    return decoded_output, decoded_offsets
