+++ deepspeech_pytorch/testing.py
@@ -1,0 +1,51 @@
+from tqdm import tqdm
+
+
+def evaluate(test_loader,
+             device,
+             model,
+             decoder,
+             target_decoder,
+             save_output=None,
+             verbose=False,
+             half=False):
+    model.eval()
+    total_cer, total_wer, num_tokens, num_chars = 0, 0, 0, 0
+    output_data = []
+    for i, (data) in tqdm(enumerate(test_loader), total=len(test_loader)):
+        inputs, targets, input_percentages, target_sizes = data
+        input_sizes = input_percentages.mul_(int(inputs.size(3))).int()
+        inputs = inputs.to(device)
+        if half:
+            inputs = inputs.half()
+        # unflatten targets
+        split_targets = []
+        offset = 0
+        for size in target_sizes:
+            split_targets.append(targets[offset:offset + size])
+            offset += size
+
+        out, output_sizes = model(inputs, input_sizes)
+
+        decoded_output, _ = decoder.decode(out, output_sizes)
+        target_strings = target_decoder.convert_to_strings(split_targets)
+
+        if save_output is not None:
+            # add output to data array, and continue
+            output_data.append((out.cpu(), output_sizes, target_strings))
+        for x in range(len(target_strings)):
+            transcript, reference = decoded_output[x][0], target_strings[x][0]
+            wer_inst = decoder.wer(transcript, reference)
+            cer_inst = decoder.cer(transcript, reference)
+            total_wer += wer_inst
+            total_cer += cer_inst
+            num_tokens += len(reference.split())
+            num_chars += len(reference.replace(' ', ''))
+            if verbose:
+                print("Ref:", reference.lower())
+                print("Hyp:", transcript.lower())
+                print("WER:", float(wer_inst) / len(reference.split()),
+                      "CER:", float(cer_inst) / len(reference.replace(' ', '')), "\n")
+    wer = float(total_wer) / num_tokens
+    cer = float(total_cer) / num_chars
+    return wer * 100, cer * 100, output_data
