+++ requirements.txt
@@ -13,4 +13,4 @@ flask
 sox
 sklearn
 soundfile
-hydra-core
+hydra-core --pre
