+++ setup.py
@@ -1,0 +1,8 @@
+from setuptools import setup, find_packages
+
+setup(name='deepspeech_pytorch',
+      version='0.1',
+      author='SeanNaren',
+      packages=find_packages(),
+      zip_safe=False,
+      )
